Subscribe meetingSeriesOverview on home-view - fixes #362
Since we need to access the meeting series on both tabs (Meeting Series List and Action Items Tab) which are hold by the home-view, I moved the subscription to the parent view. Additionally, I adjusted the `meetingSeriesOverview`-publication to pass the `availableLabels`, too. They are required to show the labels of the action items in the Action Items Tab.
