+++ client/templates/home.html
@@ -1,6 +1,7 @@
 <template name="home">
-
-    {{#if currentUser}}
+    {{#if authenticating}}
+        {{> loading}}
+    {{else}}
         <ul class="nav nav-tabs">
             <li id="tab_meetings" data-template="meetingSeriesList" role="presentation" class="{{isTabActive 'tab_meetings'}}"><a>Meetings</a></li>
             <li id="tab_actionItems" data-template="actionItemList" data-action="actionItems" role="presentation" class="{{isTabActive 'tab_actionItems'}}"><a>Action Items</a></li>
@@ -8,12 +9,6 @@
         <div class="active">
             {{> Template.dynamic template=tab}}
         </div>
-    {{else}}
-
-      <!-- This login template embedding is for security only... -->
-      <!-- Normally the router.js decides to render the login template if no user is logged in -->
-      {{> login}}
-
     {{/if}}
 
 </template>
