+++ client/templates/home.js
@@ -1,3 +1,4 @@
+import { Meteor } from 'meteor/meteor';
 import { Template } from 'meteor/templating';
 import { $ } from 'meteor/jquery';
 import { ReactiveVar } from 'meteor/reactive-var';
@@ -10,9 +11,18 @@ Template.home.created = function () {
 Template.home.onCreated(function () {
     this.activeTabTemplate = new ReactiveVar('meetingSeriesList');
     this.activeTabId = new ReactiveVar('tab_meetings');
+    this.seriesReady = new ReactiveVar();
+    this.autorun(() => {
+        this.subscribe('meetingSeriesOverview');
+        this.seriesReady.set(this.subscriptionsReady());
+    });
 });
 
 Template.home.helpers({
+    authenticating() {
+        const subscriptionReady = Template.instance().seriesReady.get();
+        return Meteor.loggingIn() || !subscriptionReady;
+    },
     isTabActive: function (tabId) {
         return (Template.instance().activeTabId.get() === tabId) ? 'active' : '';
     },
