+++ client/templates/meetingseries/meetingSeriesList.html
@@ -1,24 +1,20 @@
 <template name="meetingSeriesList">
-    {{#if authenticating}}
-        {{> loading}}
-    {{else}}
-        {{> meetingSeriesAdd}}
-        {{#if meetingSeriesAmountBiggerFour}}
-            {{> meetingSeriesSearch updateSearchQuery=updateSearchQuery}}
-        {{/if}}
-        <br>
-        <br>
-
-        <ul class="list-group" style="font-size: 130%;">
-            {{#if meetingSeriesRow}}
-                {{#each meetingSeriesRow}}
-                    {{> meetingSeriesOverview}}
-                {{/each}}
-            {{else}}
-                <span id="id_noresults">No Results for this input..</span>
-            {{/if}}
-        </ul>
-
-        {{> quickHelp context='meetingSeriesList' contentTmpl='quickHelpMeetingSeriesList'}}
+    {{> meetingSeriesAdd}}
+    {{#if meetingSeriesAmountBiggerFour}}
+        {{> meetingSeriesSearch updateSearchQuery=updateSearchQuery}}
     {{/if}}
+    <br>
+    <br>
+
+    <ul class="list-group" style="font-size: 130%;">
+        {{#if meetingSeriesRow}}
+            {{#each meetingSeriesRow}}
+                {{> meetingSeriesOverview}}
+            {{/each}}
+        {{else}}
+            <span id="id_noresults">No Results for this input..</span>
+        {{/if}}
+    </ul>
+
+    {{> quickHelp context='meetingSeriesList' contentTmpl='quickHelpMeetingSeriesList'}}
 </template>
